requirements.txt
crispy-bootstrap5==0.7
dj-database-url==1.2.0
django-crispy-forms==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
environs==9.5.0
et-xmlfile==1.1.0
fpdf==1.7.2
marshmallow==3.19.0
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
XlsxWriter==3.0.9
